Bug fix which creates too many database connection
This submission changes the mechanism of establishing database connection.
In the old version, each DAO object maintains a database connection which
is unacceptable in BlueHost's server. Because that server has a limitation
in max user connection. So, in this submission, a global database
connection is created to deal with this problem.